Convert Taxa::TaxaReferences#references to a service object
