+++ spec/services/taxa/what_links_here_spec.rb
@@ -1,23 +1,27 @@
-require 'spec_helper'
+require "spec_helper"
 
-describe Taxon do
+describe Taxa::WhatLinksHere do
+  let(:reference) { create :article_reference }
+
   describe "#references" do
     let!(:atta) { create_genus 'Atta' }
+    subject { described_class.new(atta) }
 
     it "has no references, if alone" do
-      expect(atta.references.size).to eq 0
+      expect(subject.taxon_references.size).to eq 0
     end
 
     describe "references in Taxon fields" do
       it "has a reference if it's a taxon's genus" do
         species = create_species genus: atta
-        expect(atta.references).to match_array [
+        expect(subject.taxon_references).to match_array [
           { table: 'taxa', field: :genus_id, id: species.id }
         ]
       end
 
       it "has a reference if it's a taxon's subfamily" do
-        expect(atta.subfamily.references).to match_array [
+        subject = described_class.new(atta.subfamily)
+        expect(subject.taxon_references).to match_array [
           { table: 'taxa', field: :subfamily_id, id: atta.id },
           { table: 'taxa', field: :subfamily_id, id: atta.tribe.id }
         ]
@@ -29,21 +33,21 @@ describe Taxon do
         eciton = create_genus 'Eciton'
         eciton.update_attribute :type_taxt, "{tax #{atta.id}}"
 
-        expect(atta.references).to match_array [
+        expect(subject.taxon_references).to match_array [
           { table: 'taxa', field: :type_taxt, id: eciton.id }
         ]
       end
 
       it "doesn't return references in its own taxt" do
         atta.update_attribute :type_taxt, "{tax #{atta.id}}"
-        expect(atta.references).to be_empty
+        expect(subject.taxon_references).to be_empty
       end
     end
 
     describe "Reference in its authorship taxt" do
       it "doesn't consider this an external reference" do
         atta.protonym.authorship.update_attribute :notes_taxt, "{tax #{atta.id}}"
-        expect(atta.references).to be_empty
+        expect(subject.taxon_references).to be_empty
       end
     end
 
@@ -53,10 +57,13 @@ describe Taxon do
         # Previously `eciton.become_junior_synonym_of atta`
         Synonym.create! junior_synonym: eciton, senior_synonym: atta
 
-        expect(atta.references).to match_array [
+        subject = described_class.new(atta)
+        expect(subject.taxon_references).to match_array [
           { table: 'synonyms', field: :senior_synonym_id, id: eciton.id }
         ]
-        expect(eciton.references).to match_array [
+
+        subject = described_class.new(eciton)
+        expect(subject.taxon_references).to match_array [
           { table: 'synonyms', field: :junior_synonym_id, id: atta.id }
         ]
       end
@@ -65,6 +72,7 @@ describe Taxon do
 
   describe "non-taxt references" do
     let!(:atta) { create_genus 'Atta' }
+    subject { described_class.new(atta) }
 
     context "taxon has non-taxt references" do
       let!(:eciton) { create_genus 'Eciton' }
@@ -75,24 +83,24 @@ describe Taxon do
 
       describe "#nontaxt_references" do
         it "returns only non-taxt references" do
-          expect(atta.nontaxt_references).to match_array [
+          expect(subject.nontaxt_references).to match_array [
             { table: 'taxa', field: :homonym_replaced_by_id, id: eciton.id }
           ]
         end
       end
 
       describe "#any_nontaxt_references?" do
-        it { expect(atta.any_nontaxt_references?).to be true }
+        it { expect(subject.any_nontaxt_references?).to be true }
       end
     end
 
     context "taxon has no non-taxt references" do
       describe "#nontaxt_references" do
-        it { expect(atta.nontaxt_references).to be_empty }
+        it { expect(subject.nontaxt_references).to be_empty }
       end
 
       describe "#any_nontaxt_references?" do
-        it { expect(atta.any_nontaxt_references?).to be_falsey }
+        it { expect(subject.any_nontaxt_references?).to be_falsey }
       end
     end
   end
